Change states to controllers
