+++ Assets/Scripts/Minion/Controllers/MinionIdleController.cs.meta
