--- Assets/Scripts/Minion/States/MinionChargeAttackController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2ab10fd0c2aefc478e274736e2bc442
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
