--- Assets/Scripts/Minion/States/MinionController.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections;
-using FSM;
-using UnityEngine;
-
-namespace Minion.States
-{
-    public abstract class MinionController : MonoBehaviour
-    {
-        [HideInInspector] public GameObject target;
-
-        public abstract void Enter();
-        public abstract void OnUpdate();
-        public abstract void Exit();
-    }
-}
